requirements.txt
pytorch
selenium
pandas
numpy
pillow
torchvision
matplotlib==3.4.3
split-folders
opencv-python
urllib3
deep-translator
scikit-learn
random2
time
pickle
flask
seaborn
mediapipe


